emltriage/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "emltriage"
version = "0.1.0"
description = "DFIR-grade email analysis tool with deterministic extraction and AI narrative"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "DFIR Team"}
]
keywords = ["dfir", "email", "forensics", "security", "eml", "threat-intel"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Topic :: Communications :: Email",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-magic>=0.4.27",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "mail-parser>=3.15.0",
    "dnspython>=2.4.0",
    "dkimpy>=1.1.0",
    "oletools>=0.60",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "email-validator>=2.1.0",
    "structlog>=24.1.0",
    "xxhash>=3.4.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
    "anthropic>=0.8.0",
    "ollama>=0.1.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "types-python-dateutil",
    "types-beautifulsoup4",
]

[project.scripts]
emltriage = "emltriage.cli:app"

[project.urls]
Homepage = "https://github.com/dfir/emltriage"
Repository = "https://github.com/dfir/emltriage"
Documentation = "https://github.com/dfir/emltriage/blob/main/README.md"

[tool.hatch.build.targets.wheel]
packages = ["emltriage"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "oletools.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dkim.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "magic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mail_parser.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["emltriage"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
